% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{check_fetch_most_recent_obstruction_file}
\alias{check_fetch_most_recent_obstruction_file}
\title{check to see if the proposed zip file is the most recent}
\usage{
check_fetch_most_recent_obstruction_file(dir = ".", proposed_zip = NULL)
}
\description{
check to see if the proposed zip file is the most recent
}
