% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suitability_generator.R
\name{explanatory_vars_to_rdata_file}
\alias{explanatory_vars_to_rdata_file}
\title{generate a spatially-consistent stack of explanatory variables and cache the results to disk for
redistribution}
\usage{
explanatory_vars_to_rdata_file(rasters = NULL, filename = NULL, n = NULL)
}
\description{
generate a spatially-consistent stack of explanatory variables and cache the results to disk for
redistribution
}
