% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suitability_generator.R
\name{extractDensities}
\alias{extractDensities}
\title{extract quantiles from a continuous raster surface as descrete spatial polygons}
\usage{
extractDensities(x, p = c(50, 95), use_smoother = 0)
}
\description{
extract quantiles from a continuous raster surface as descrete spatial polygons
}
