% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{fit_boosted_gam}
\alias{fit_boosted_gam}
\title{fit a generalized additive model to a user-specified dataset using an optional functional specification}
\usage{
fit_boosted_gam(formula = NULL, training_data = NULL, vars = NULL,
  control = mboost::boost_control(center = T))
}
\description{
fit a generalized additive model to a user-specified dataset using an optional functional specification
}
