% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{fit_maxent}
\alias{fit_maxent}
\title{fit a maxent model to the input dataset}
\usage{
fit_maxent(formula = NULL, training_data = NULL, predictor_stack = NULL)
}
\description{
fit a maxent model to the input dataset
}
