% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suitability_generator.R
\name{gen_gam_suitability_raster}
\alias{gen_gam_suitability_raster}
\title{function that will merge presences and absences SpatialPoints data.frame using a 'year' attribute}
\usage{
gen_gam_suitability_raster(m = NULL, explanatory_vars = NULL,
  write = NULL, quietly = F, normalize = T, n = NULL)
}
\description{
function that will merge presences and absences SpatialPoints data.frame using a 'year' attribute
}
