% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{gen_pseudo_absences}
\alias{gen_pseudo_absences}
\title{generate a large pool of random pseudo-absences within a geographic boundary
using a set of user-specified wind turbine point locations. The pool will be
arbitrarily large (nrow(pts)*iter) with the intention that the dataset will
be downsampled through a bagging procedure down-the-line.}
\usage{
gen_pseudo_absences(pts = NULL, boundary = NULL, iter = 5,
  buffer_width = 1000, quietly = T)
}
\description{
generate a large pool of random pseudo-absences within a geographic boundary
using a set of user-specified wind turbine point locations. The pool will be
arbitrarily large (nrow(pts)*iter) with the intention that the dataset will
be downsampled through a bagging procedure down-the-line.
}
