% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suitability_generator.R
\name{gen_rf_suitability_raster}
\alias{gen_rf_suitability_raster}
\title{function that will generate a suitability surface using a random forest model}
\usage{
gen_rf_suitability_raster(m = NULL, explanatory_vars = NULL, write = NULL)
}
\description{
function that will generate a suitability surface using a random forest model
}
