% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{generate_uniform_grid}
\alias{generate_uniform_grid}
\title{Generates a uniform grid of polygons such that each grid 'cell' is
a user-specified size. This follows the original methodology of
A. Daniels (2017) used for national wildlife refuge planning (LCD).}
\usage{
generate_uniform_grid(boundary = NULL, cell_size = 9397.127)
}
\description{
Generates a uniform grid of polygons such that each grid 'cell' is
a user-specified size. This follows the original methodology of
A. Daniels (2017) used for national wildlife refuge planning (LCD).
}
