% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suitability_generator.R
\name{min_max_normalize}
\alias{min_max_normalize}
\title{hidden function that will perform a min-max normalization on an input raster to ensure it is projected as 0-to-1}
\usage{
min_max_normalize(d)
}
\description{
hidden function that will perform a min-max normalization on an input raster to ensure it is projected as 0-to-1
}
