% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{scrape_turbine_buildout}
\alias{scrape_turbine_buildout}
\title{Downloads a managed CSV file that contains projections of wind
energy build-out and extracts 1.) total area of build-out and
2.) number of turbines associated with WITO projections}
\usage{
scrape_turbine_buildout()
}
\description{
Downloads a managed CSV file that contains projections of wind
energy build-out and extracts 1.) total area of build-out and
2.) number of turbines associated with WITO projections
}
