% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suitability_generator.R
\name{smoother}
\alias{smoother}
\title{short-hand guassian smoother function to smooth out slivers --
useful for feature extraction from a noisy raster surface (e.g.,
from random forests)}
\usage{
smoother(r = NULL, smoothing_fun = mean, width = 33, progress = "text",
  ...)
}
\description{
short-hand guassian smoother function to smooth out slivers --
useful for feature extraction from a noisy raster surface (e.g.,
from random forests)
}
