% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{merge_presences_absences_by_year}
\alias{merge_presences_absences_by_year}
\title{merge presences and absences SpatialPoints data.frame using a 'year' attribute}
\usage{
merge_presences_absences_by_year(presences = NULL, absences = NULL,
  years = NULL, bag = T)
}
\description{
merge presences and absences SpatialPoints data.frame using a 'year' attribute
}
